bindings/python: fix UTF-8 handling for Python 3
The Python bindings for Python 2 stored the attributes as UTF-8 strings
and in the local encoding and preferred the local encoding.

Since UTF-8 strings are now the default in Python 3, the Python bindings
now assume UTF-8 strings in their high level frontend but need a byte
string for the C side. All encoding conversion is done on the Python
side.